api/pyproject.toml
[project]
name = "bedrock-model-compare"
version = "0.1.0"
description = "A tool to execute and compare AWS Bedrock models in parallel"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.34.0",
    "botocore[crt]>=1.34.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
bedrock-executor = "bedrock_executor:main"
bedrock-server = "api_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = []
